Handle bash-style variables in build settings
Bash-style variables defined in build settings are incorrectly parsed as
literal values, causing settings like `${PRODUCT_NAME}` to be escaped
and returned as `__PRODUCT_NAME_`.
